+++ src/main/java/com/riwi/riwi_ecomerce/domain/entities/ShoppingCart.java
@@ -1,0 +1,33 @@
+package com.riwi.riwi_ecomerce.domain.entities;
+
+import java.util.List;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Entity(name = "shopping_cart")
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ShoppingCart {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private UserEntity user;
+
+    @ToString.Exclude
+    @EqualsAndHashCode.Exclude
+    private List<ProductEntity> products;
+}
