+++ src/main/java/com/riwi/riwi_ecomerce/domain/entities/UserEntity.java
@@ -1,7 +1,7 @@
-package com.riwi_ecomerce.riwi_ecomerce.domain.entities;
+package com.riwi.riwi_ecomerce.domain.entities;
 
-import com.riwi_ecomerce.riwi_ecomerce.util.enums.RoleUser;
-import com.riwi_ecomerce.riwi_ecomerce.util.enums.StatusUser;
+import com.riwi.riwi_ecomerce.util.enums.RoleUser;
+import com.riwi.riwi_ecomerce.util.enums.StatusUser;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
