+++ src/main/java/com/riwi/riwi_ecomerce/util/enums/RoleUser.java
@@ -1,4 +1,4 @@
-package com.riwi_ecomerce.riwi_ecomerce.util.enums;
+package com.riwi.riwi_ecomerce.util.enums;
 
 public enum RoleUser {
     CODER,
