--- src/main/java/com/riwi_ecomerce/riwi_ecomerce/domain/entities/ProductEntity.java
@@ -1,35 +1,0 @@
-package com.riwi_ecomerce.riwi_ecomerce.domain.entities;
-
-import java.math.BigDecimal;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity(name = "product")
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ProductEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(nullable = false)
-    private String name;
-    private String description;
-    @Column(nullable = false)
-    private BigDecimal price;
-    @Column(nullable = false)
-    private String imageUrl;
-    @Column(nullable = false)
-    private Integer stock;
-    @Column(nullable = false)
-    private String category;
-}
